+++ ui/app/src/components/PathNavigatorTree/PathNavigatorTree.tsx
@@ -61,6 +61,9 @@ import SystemType from '../../models/SystemType';
 import { PathNavigatorTreeItemProps } from './PathNavigatorTreeItem';
 import { UNDEFINED } from '../../utils/constants';
 import SimpleAjaxError from '../../models/SimpleAjaxError';
+import usePossibleTranslation from '../../hooks/usePossibleTranslation';
+import TranslationOrText from '../../models/TranslationOrText';
+import { useIntl } from 'react-intl';
 
 export interface PathNavigatorTreeProps
   extends Pick<
@@ -68,7 +71,7 @@ export interface PathNavigatorTreeProps
     'showNavigableAsLinks' | 'showPublishingTarget' | 'showWorkflowState' | 'showItemMenu'
   > {
   id: string;
-  label: string;
+  label: TranslationOrText;
   rootPath: string;
   sortStrategy?: GetChildrenOptions['sortStrategy'];
   order?: GetChildrenOptions['order'];
@@ -129,17 +132,18 @@ interface Menu {
 // };
 
 export function PathNavigatorTree(props: PathNavigatorTreeProps) {
+  const { formatMessage } = useIntl();
+  const translatedLabel = usePossibleTranslation(props.label) || formatMessage({ defaultMessage: '(No name)' });
   // region const { ... } = props;
   const {
-    label,
-    id = props.label.replace(/\s/g, ''),
+    rootPath,
+    id = rootPath,
     excludes,
     limit = 10,
     icon,
     expandedIcon,
     collapsedIcon,
     container,
-    rootPath,
     initialExpanded,
     initialCollapsed = true,
     collapsible = true,
@@ -334,7 +338,7 @@ export function PathNavigatorTree(props: PathNavigatorTreeProps) {
     <>
       <PathNavigatorTreeUI
         classes={{ header: classes?.header }}
-        title={label}
+        title={translatedLabel}
         active={active}
         icon={expandedIcon && collapsedIcon ? (state.collapsed ? collapsedIcon : expandedIcon) : icon}
         container={container}