+++ ui/app/src/components/PublishOnDemandWidget/PublishOnDemandWidget.tsx
@@ -489,7 +489,7 @@ export function PublishOnDemandWidget(props: PublishOnDemandWidgetProps) {
                               defaultMessage="Publish changes made in Studio via the UI"
                             />
                           }
-                          secondary="By path"
+                          secondary={<FormattedMessage defaultMessage="By path" />}
                         />
                       }
                       className={classes.byPathModeSelector}
@@ -508,7 +508,7 @@ export function PublishOnDemandWidget(props: PublishOnDemandWidgetProps) {
                               defaultMessage="Publish changes made via direct git actions against the repository or pulled from a remote repository"
                             />
                           }
-                          secondary="By tags or commit ids"
+                          secondary={<FormattedMessage defaultMessage="By tags or commit ids" />}
                         />
                       }
                     />
@@ -526,7 +526,9 @@ export function PublishOnDemandWidget(props: PublishOnDemandWidgetProps) {
                               defaultMessage="Publish everything"
                             />
                           }
-                          secondary="Publish all changes on the repo to the publishing target you choose"
+                          secondary={
+                            <FormattedMessage defaultMessage="Publish all changes on the repo to the publishing target you choose" />
+                          }
                         />
                       }
                     />