+++ ui/guest/scripts/package.js
@@ -34,7 +34,6 @@ const peerDeps = [
   '@mui/styles',
   'react',
   'react-dom',
-  'react-intl',
   'react-redux',
   'react-router-dom',
   'redux',